+++ src/main/java/org/lagonette/hellos/bean/helloasso/HelloAssoPaymentStateEnum.java
@@ -1,0 +1,5 @@
+package org.lagonette.hellos.bean.helloasso;
+
+public enum HelloAssoPaymentStateEnum {
+    Pending, Authorized, Refused, Unknown, Registered, Error, Refunded, Refunding, Waiting, Canceled, Contested, WaitingBankValidation, WaitingBankWithdraw
+}